clientes_api/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
psycopg-binary==3.1.11
Flask-Migrate==4.0.5
WTForms==3.0.1
Flask-WTF==1.1.1
passlib==1.7.4
bcrypt==4.0.1
Flask-JWT-Extended==4.5.3